the philos are dying properly, but it's not exiting sometimes

+++ philo.h
@@ -6,7 +6,7 @@
 /*   By: msoklova <msoklova@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/02 14:29:50 by msoklova          #+#    #+#             */
-/*   Updated: 2024/11/20 16:34:54 by msoklova         ###   ########.fr       */
+/*   Updated: 2024/11/20 19:24:12 by msoklova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ void		*death_monitor(void *arg);
 
 /* UTILS */
 int				ft_atoi(const char *str);
-long long		curr_time(void);
+long			curr_time(void);
 void			print_action(t_events *events, int id, const char *action);
 int				ft_usleep(useconds_t time);
